dora/requirements.txt
PyGithub==2.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.31.0
